refactor: extract injectAll functions to their own file
